+++ packages/utils/src/datatypes/ChangeLog.ts
@@ -2,6 +2,7 @@ import { BaseObject } from './BaseObject';
 import { Changes } from './Card';
 
 export interface CubeChangeLog {
+  id: string;
   cubeId: string;
   date: number;
   changelog: Changes;